I have refactored the code of the library which is now much simpler and easier to read and (hopefully) understand.
In addition, I found a new (tiny but technical) bug: edges
characterized with attributes pointing to a block of multiple vertices
only affected the first vertex of the block, but not the others. It
has been fixed.

Finally, two new Google unit tests have been added for testing that
edges are properly parsed along with their attributes in the presence
of blocks with multiple vertices and paths of an arbitrary length.